Edits on tests.
